Extract Allhomes past sales data for a single division ID and year.

Extract Allhomes past sales data for a division ID and year. The division ID is the ID for a specific division (suburb). This is an internal function that gets called by vectorised get_past_sales_data(). This function is currently exported but this may change in the future.

extract_past_sales_data(division_id, year, quiet = FALSE)

Arguments

division_id

An integer scalar denoting the Allhomes division ID.

year

An integer scalar denoting the year.

quiet

If TRUE then messages are suppressed. Currently ignored.

Value

A tibble.

Examples

# \donttest{
extract_past_sales_data(14743, 2021)
#> [2022-09-13 06:08:57] Finding data for ID=14743, year=2021...
#> [2022-09-13 06:08:57] URL: https://www.allhomes.com.au/ah/research/_/121474312/sale-history?year=2021
#> [2022-09-13 06:09:00] Found 106 entries.
#>                    address bedrooms bathrooms ensuites garages carports
#> 1       25 Corroboree Park       NA        NA       NA      NA       NA
#> 2        24 Raymond Street        4         2       NA       2        0
#> 3      18 Higgins Crescent       NA        NA       NA      NA       NA
#> 4      32 Officer Crescent        3         3       NA       1       NA
#> 5     94 Wakefield Gardens       NA        NA       NA      NA       NA
#> 6       49 Campbell Street       NA        NA       NA      NA       NA
#> 7          26 Angas Street        4         1       NA       2        0
#> 8         240 Duffy Street       NA        NA       NA      NA       NA
#> 9         105 Duffy Street        4         2       NA       1        0
#> 10        10 Fisher Street        3         2       NA       1        0
#> 11       62 Raymond Street        4         3       NA       2        0
#> 12        104 Duffy Street        4         5       NA      NA       NA
#> 13          1 Dibbs Street        4         2       NA       2        0
#> 14        264 Duffy Street        6         2       NA       1        0
#> 15          7 Duffy Street       NA        NA       NA      NA       NA
#> 16       44 Foveaux Street       NA        NA       NA      NA       NA
#> 17          5 Ebden Street       NA        NA       NA      NA       NA
#> 18      31 Chisholm Street        0         0       NA       0        0
#> 19       1/85 Duffy Street       NA        NA       NA      NA       NA
#> 20         83 Duffy Street        4         2       NA       0        0
#> 21       34 Raymond Street        3         1       NA       0        2
#> 22        139 Duffy Street        6         4       NA       2        0
#> 23     62 Limestone Avenue       NA        NA       NA      NA       NA
#> 24    26 Sherbrooke Street        4         2       NA       0        2
#> 25        45 Cowper Street       NA        NA       NA      NA       NA
#> 26     17 Davenport Street        3         1       NA       0        2
#> 27    17A Davenport Street       NA        NA       NA      NA       NA
#> 28         9 Bonney Street        4         2       NA       2        0
#> 29     2/21A Cowper Street       NA        NA       NA      NA       NA
#> 30        198 Duffy Street        4         2       NA       2        0
#> 31      26 Paterson Street       NA        NA       NA      NA       NA
#> 32         13 Duffy Street        3         1       NA       2        2
#> 33          4 Agnew Street       NA        NA       NA      NA       NA
#> 34      45 Campbell Street       NA        NA       NA      NA       NA
#> 35        15 Howe Crescent       NA        NA       NA      NA       NA
#> 36         10 Ebden Street        3         2       NA       1        0
#> 37        4 Hoddle Gardens        3         1       NA       2        0
#> 38          3 Piper Street       NA        NA       NA      NA       NA
#> 39      83 Hannan Crescent       NA        NA       NA      NA       NA
#> 40       32 Canning Street       NA        NA       NA      NA       NA
#> 41        36A Duffy Street        3         2       NA       3        0
#> 42       2/36 Duffy Street       NA        NA       NA      NA       NA
#> 43         19 Duffy Street       NA        NA       NA      NA       NA
#> 44         53 Tyson Street       NA        NA       NA      NA       NA
#> 45         93 Ebden Street       NA        NA       NA      NA       NA
#> 46         7 Gillen Street        2         2       NA       1        0
#> 47         92 Duffy Street        3         1       NA       1        0
#> 48    3/10 Paterson Street        3         2       NA       2        0
#> 49   19 Shortland Crescent       NA        NA       NA      NA       NA
#> 50    102 Limestone Avenue       NA        NA       NA      NA       NA
#> 51         11 Tyson Street       NA        NA       NA      NA       NA
#> 52         14 Lalor Street        3         1       NA       1        0
#> 53       3 Hannan Crescent        3         1       NA       0        0
#> 54     54 Limestone Avenue        3         2       NA       1        0
#> 55         2A Ebden Street        0         0       NA       0        0
#> 56     2/6 Chisholm Street       NA        NA       NA      NA       NA
#> 57  19 Fairbridge Crescent       NA        NA       NA      NA       NA
#> 58    1/31A Raymond Street       NA        NA       NA      NA       NA
#> 59      31A Raymond Street        3         2       NA       2        0
#> 60      43 Campbell Street       NA        NA       NA      NA       NA
#> 61    2/15 Chisholm Street       NA        NA       NA      NA       NA
#> 62    51 Sherbrooke Street        3         1       NA      NA       NA
#> 63        187 Duffy Street       NA        NA       NA      NA       NA
#> 64      32 Paterson Street       NA        NA       NA      NA       NA
#> 65        40 Hawdon Street        3         1       NA       0        0
#> 66      50 Hannan Crescent        5         3       NA       3        0
#> 67        11 Gillen Street        3         1       NA       2        0
#> 68     75 Officer Crescent       NA        NA       NA      NA       NA
#> 69         22 Angas Street        5         2       NA       2        0
#> 70        10 Fisher Street       NA        NA       NA      NA       NA
#> 71         35 Ebden Street       NA        NA       NA      NA       NA
#> 72     30 O'Connell Street        3         1       NA       1        0
#> 73         21 Agnew Street       NA        NA       NA      NA       NA
#> 74       182 Majura Avenue        3         1       NA       1        0
#> 75        136 Duffy Street       NA        NA       NA      NA       NA
#> 76      72 Hannan Crescent       NA        NA       NA      NA       NA
#> 77        10 Gillen Street       NA        NA       NA      NA       NA
#> 78       24 Calvert Street        3         1       NA       1       NA
#> 79    61 Sherbrooke Street       NA        NA       NA      NA       NA
#> 80       10 Calvert Street        3         1       NA       1        2
#> 81        74 Majura Avenue       NA        NA       NA      NA       NA
#> 82       1/6 Cowper Street       NA        NA       NA      NA       NA
#> 83         6 Cowper Street        2         1       NA       1        0
#> 84     26 O'Connell Street        3         1       NA       0        0
#> 85    75 Sherbrooke Street       NA        NA       NA      NA       NA
#> 86         21 Piper Street        2         1       NA      NA       NA
#> 87    96 Wakefield Gardens       NA        NA       NA      NA       NA
#> 88        2/6 Hobson Place       NA        NA       NA      NA       NA
#> 89       196 Majura Avenue        3         2       NA       2        0
#> 90     2/196 Majura Avenue       NA        NA       NA      NA       NA
#> 91        111 Hannan Place       NA        NA       NA      NA       NA
#> 92       9 Winnecke Street       NA        NA       NA      NA       NA
#> 93       2/95 Ebden Street        2         2       NA       0        1
#> 94           6 Bodel Place       NA        NA       NA      NA       NA
#> 95      18 Campbell Street       NA        NA       NA      NA       NA
#> 96     90 Limestone Avenue       NA        NA       NA      NA       NA
#> 97        13 Hayley Street       NA        NA       NA      NA       NA
#> 98                1 Street       NA        NA       NA      NA       NA
#> 99        56 Majura Avenue       NA        NA       NA      NA       NA
#> 100         4 Agnew Street       NA        NA       NA      NA       NA
#> 101     6 O'Connell Street       NA        NA       NA      NA       NA
#> 102         5 Ebden Street       NA        NA       NA      NA       NA
#> 103        26 Angas Street       NA        NA       NA      NA       NA
#> 104      2/6B Alt Crescent       NA        NA       NA      NA       NA
#> 105    62 Limestone Avenue       NA        NA       NA      NA       NA
#> 106     36 Paterson Street       NA        NA       NA      NA       NA
#>     contract_date transfer_date  list_date   price block_size transfer_type
#> 1      13/11/2021    21/02/2022       <NA> 2775000       1143 Full transfer
#> 2      25/09/2021    01/12/2021 04/09/2021 2440000        759 Full transfer
#> 3      24/07/2021    13/09/2021       <NA> 2400000       1176 Full transfer
#> 4      23/12/2021    01/04/2022 20/09/2021 2350000        839 Full transfer
#> 5      04/03/2021    30/04/2021       <NA> 2220000        806 Full transfer
#> 6      12/11/2021    03/02/2022       <NA> 2200000        947 Full transfer
#> 7      30/10/2021    21/01/2022 29/09/2021 2170000       1099 Full transfer
#> 8      29/09/2021    14/01/2022       <NA> 2156000       1085 Full transfer
#> 9      21/08/2021    20/10/2021 28/07/2021 2110000        952 Full transfer
#> 10     06/11/2021    07/01/2022 21/10/2021 1960000        725 Full transfer
#> 11     10/04/2021    20/07/2021 18/03/2021 1955000        564 Full transfer
#> 12     27/03/2021    22/07/2021 27/02/2021 1950000       1188 Full transfer
#> 13     30/03/2021    24/06/2021 16/02/2021 1885000        711 Full transfer
#> 14     18/12/2021    04/03/2022 03/12/2021 1850000        910 Full transfer
#> 15     12/03/2021    13/07/2021       <NA> 1830000        867 Full transfer
#> 16     24/07/2021    06/09/2021       <NA> 1811000        953 Full transfer
#> 17     13/11/2021    09/02/2022       <NA> 1755000        903 Full transfer
#> 18     04/05/2021          <NA> 09/04/2021 1751000        861 Full transfer
#> 19     19/11/2021    10/01/2022       <NA> 1750000       1294 Full transfer
#> 20     19/11/2021          <NA> 07/10/2021 1750000          0 Full transfer
#> 21     22/12/2021    15/03/2022 25/11/2021 1750000       1011 Full transfer
#> 22     03/03/2021    06/08/2021 03/03/2021 1735000       1160 Full transfer
#> 23     27/05/2021    13/09/2021       <NA> 1720000       1102 Full transfer
#> 24     17/04/2021    14/07/2021 25/03/2021 1710000        832 Full transfer
#> 25     24/09/2021    09/12/2021       <NA> 1702000       1134 Full transfer
#> 26     24/07/2021          <NA> 02/07/2021 1695000          0 Full transfer
#> 27     24/07/2021    25/10/2021       <NA> 1695000       1165 Full transfer
#> 28     27/02/2021    18/05/2021 03/02/2021 1691000       1054 Full transfer
#> 29     17/07/2021    24/09/2021       <NA> 1690000       1140 Full transfer
#> 30     09/06/2021    29/07/2021 20/05/2021 1685000       1035 Full transfer
#> 31     13/03/2021    17/05/2021       <NA> 1680000        697 Full transfer
#> 32     04/12/2021    03/02/2022 11/11/2021 1675000        844 Full transfer
#> 33     28/05/2021    08/12/2021       <NA> 1670000        706 Full transfer
#> 34     10/09/2021    28/10/2021       <NA> 1660000        982 Full transfer
#> 35     10/02/2021    30/04/2021       <NA> 1650000        830 Full transfer
#> 36     28/01/2021    01/04/2021 11/01/2021 1621000        831 Full transfer
#> 37     25/09/2021    09/12/2021 04/08/2021 1610000        743 Full transfer
#> 38     23/10/2021    17/12/2021       <NA> 1600000        661 Full transfer
#> 39     21/05/2021    22/07/2021       <NA> 1590000        694 Full transfer
#> 40     11/03/2021    25/05/2021       <NA> 1570000        826 Full transfer
#> 41     04/02/2021          <NA> 26/11/2020 1560000        404 Full transfer
#> 42     04/02/2021    23/04/2021       <NA> 1560000        809 Full transfer
#> 43     19/02/2021    23/06/2021       <NA> 1550000        867 Full transfer
#> 44     19/06/2021    15/09/2021       <NA> 1515000       1281 Full transfer
#> 45     24/02/2021    22/04/2021       <NA> 1485000        836 Full transfer
#> 46     19/01/2021    08/04/2021 27/11/2020 1485000        641 Full transfer
#> 47     13/03/2021    26/05/2021 18/02/2021 1460000        689 Full transfer
#> 48     15/12/2021    25/01/2022 24/11/2021 1455000       4877 Full transfer
#> 49     17/04/2021    13/07/2021       <NA> 1436000        725 Full transfer
#> 50     16/06/2021    24/08/2021       <NA> 1427000       1040 Full transfer
#> 51     16/10/2021    17/12/2021       <NA> 1405000        988 Full transfer
#> 52     16/10/2021    16/12/2021 01/10/2021 1405000        562 Full transfer
#> 53     15/12/2021    02/03/2022 25/11/2021 1400000        613 Full transfer
#> 54     14/04/2021    15/07/2021 17/03/2021 1400000       1508 Full transfer
#> 55     04/05/2021          <NA> 09/04/2021 1385000        720 Full transfer
#> 56     24/07/2021    03/11/2021       <NA> 1372500       1301 Full transfer
#> 57     24/06/2021    02/09/2021       <NA> 1355000       1126 Full transfer
#> 58     23/09/2021    16/12/2021       <NA> 1335000        906 Full transfer
#> 59     23/09/2021          <NA> 01/09/2021 1335000        439 Full transfer
#> 60     14/04/2021    16/07/2021       <NA> 1330000        808 Full transfer
#> 61     08/05/2021    27/07/2021       <NA> 1322000       1114 Full transfer
#> 62     28/05/2021    03/08/2021 06/05/2021 1320000        721 Full transfer
#> 63     23/06/2021    30/08/2021       <NA> 1310000        877 Full transfer
#> 64     24/02/2021    04/06/2021       <NA> 1305000        697 Full transfer
#> 65     13/03/2021    21/05/2021 24/02/2021 1301000        711 Full transfer
#> 66     02/10/2021    01/12/2021 21/09/2021 1300000        557 Full transfer
#> 67     28/05/2021    23/07/2021 14/05/2021 1286000        757 Full transfer
#> 68     27/03/2021    26/05/2021       <NA> 1271000        654 Full transfer
#> 69     13/01/2021    07/04/2021 23/10/2020 1250000       1054 Full transfer
#> 70     04/06/2021    15/10/2021       <NA> 1250000        725 Full transfer
#> 71     08/02/2021    22/03/2021       <NA> 1225000        801 Full transfer
#> 72     19/05/2021    23/07/2021 05/05/2021 1220000        837 Full transfer
#> 73     14/04/2021    23/08/2021       <NA> 1220000        729 Full transfer
#> 74     13/11/2021    17/03/2022 25/10/2021 1211000        664 Full transfer
#> 75     27/03/2021    22/06/2021       <NA> 1205000        835 Full transfer
#> 76     22/05/2021    20/09/2021       <NA> 1200000        654 Full transfer
#> 77     08/05/2021    15/07/2021       <NA> 1200000        705 Full transfer
#> 78     13/03/2021    07/06/2021 16/02/2021 1183000        660 Full transfer
#> 79     31/03/2021    04/06/2021       <NA> 1175000        656 Full transfer
#> 80     30/01/2021    17/03/2021 07/01/2021 1165000        867 Full transfer
#> 81     14/04/2021    10/08/2021       <NA> 1151200        464 Full transfer
#> 82     27/02/2021    19/05/2021       <NA> 1126000        904 Full transfer
#> 83     27/02/2021          <NA> 04/02/2021 1126000        655 Full transfer
#> 84     31/03/2021    11/06/2021 03/03/2021 1125000        834 Full transfer
#> 85     10/02/2021    03/05/2021       <NA> 1120000        595 Full transfer
#> 86     25/05/2021    24/08/2021 06/05/2021 1100000        565 Full transfer
#> 87     04/03/2021    19/07/2021       <NA> 1035000        892 Full transfer
#> 88     24/06/2021    16/09/2021       <NA> 1025000       1090 Full transfer
#> 89     13/03/2021          <NA> 18/02/2021  950000          0 Full transfer
#> 90     13/03/2021    18/05/2021       <NA>  950000        915 Full transfer
#> 91     11/11/2021    24/11/2021       <NA>  900000        716 Full transfer
#> 92     23/06/2021    17/08/2021       <NA>  890000        572 Full transfer
#> 93     18/02/2021    15/03/2021 27/01/2021  735000       1103 Full transfer
#> 94     25/03/2021    04/06/2021       <NA>  720000        284 Full transfer
#> 95     03/06/2021    19/07/2021       <NA>  635000        811 Full transfer
#> 96     04/03/2021    30/03/2021       <NA>  575000       1189 Full transfer
#> 97     30/03/2021    28/06/2021       <NA>  450000        725 Full transfer
#> 98     28/07/2021          <NA> 11/01/2021  250000         NA Full transfer
#> 99     15/08/2021    07/02/2022       <NA>       0        516 Full transfer
#> 100    30/01/2021    21/05/2021       <NA>       0        706 Full transfer
#> 101    17/11/2021    16/03/2022       <NA>       0       1017 Full transfer
#> 102    23/05/2021    30/11/2021       <NA>       0        903 Full transfer
#> 103    01/06/2021    29/11/2021       <NA>       0       1099 Full transfer
#> 104    03/05/2021    20/10/2021       <NA>       0       1544 Full transfer
#> 105    17/01/2021    27/07/2021       <NA>       0       1102 Full transfer
#> 106    25/10/2021    02/03/2022       <NA>       0        697 Full transfer
#>     full_sale_price days_on_market sale_type sale_record_source building_size
#> 1           2775000             NA        NA  Government Record            NA
#> 2           2440000             21        NA    Combined Record            NA
#> 3           2400000             NA        NA  Government Record            NA
#> 4           2350000             94        NA    Combined Record           160
#> 5           2220000             NA        NA  Government Record            NA
#> 6           2200000             NA        NA  Government Record            NA
#> 7           2170000             31        NA    Combined Record            NA
#> 8           2156000             NA        NA  Government Record            NA
#> 9           2110000             24        NA    Combined Record            NA
#> 10          1960000             16        NA    Combined Record            NA
#> 11          1955000             23        NA    Combined Record            NA
#> 12          1950000             28        NA    Combined Record            NA
#> 13          1885000             42        NA    Combined Record            NA
#> 14          1850000             15        NA    Combined Record            NA
#> 15          1830000             NA        NA  Government Record            NA
#> 16          1811000             NA        NA  Government Record            NA
#> 17          1755000             NA        NA  Government Record            NA
#> 18          1751000             25        NA    Allhomes Record            NA
#> 19          1750000             NA        NA  Government Record            NA
#> 20          1750000             43        NA    Allhomes Record            NA
#> 21          1750000             27        NA    Combined Record            NA
#> 22          1735000             NA        NA    Combined Record            NA
#> 23          1720000             NA        NA  Government Record            NA
#> 24          1710000             23        NA    Combined Record            NA
#> 25          1702000             NA        NA  Government Record            NA
#> 26          1695000             22        NA    Allhomes Record            NA
#> 27          1695000             NA        NA  Government Record            NA
#> 28          1691000             24        NA    Combined Record            NA
#> 29          1690000             NA        NA  Government Record            NA
#> 30          1685000             20        NA    Combined Record            NA
#> 31          1680000             NA        NA  Government Record            NA
#> 32          1675000             23        NA    Combined Record            NA
#> 33          1670000             NA        NA  Government Record            NA
#> 34          1660000             NA        NA  Government Record            NA
#> 35          1650000             NA        NA  Government Record            NA
#> 36          1621000             17        NA    Combined Record            NA
#> 37          1610000             52        NA    Combined Record            NA
#> 38          1600000             NA        NA  Government Record            NA
#> 39          1590000             NA        NA  Government Record            NA
#> 40          1570000             NA        NA  Government Record            NA
#> 41          1560000             70        NA    Allhomes Record            NA
#> 42          1560000             NA        NA  Government Record            NA
#> 43          1550000             NA        NA  Government Record            NA
#> 44          1515000             NA        NA  Government Record            NA
#> 45          1485000             NA        NA  Government Record            NA
#> 46          1485000             53        NA    Combined Record            NA
#> 47          1460000             23        NA    Combined Record            NA
#> 48          1455000             21        NA    Combined Record            NA
#> 49          1436000             NA        NA  Government Record            NA
#> 50          1427000             NA        NA  Government Record            NA
#> 51          1405000             NA        NA  Government Record            NA
#> 52          1405000             15        NA    Combined Record            NA
#> 53          1400000             20        NA    Combined Record            NA
#> 54          1400000             28        NA    Combined Record            NA
#> 55          1385000             25        NA    Allhomes Record            NA
#> 56          1372500             NA        NA  Government Record            NA
#> 57          1355000             NA        NA  Government Record            NA
#> 58          1335000             NA        NA  Government Record            NA
#> 59          1335000             22        NA    Allhomes Record            NA
#> 60          1330000             NA        NA  Government Record            NA
#> 61          1322000             NA        NA  Government Record            NA
#> 62          1320000             22        NA    Combined Record            NA
#> 63          1310000             NA        NA  Government Record            NA
#> 64          1305000             NA        NA  Government Record            NA
#> 65          1301000             17        NA    Combined Record            NA
#> 66          1300000             11        NA    Combined Record            NA
#> 67          1286000             14        NA    Combined Record            NA
#> 68          1271000             NA        NA  Government Record            NA
#> 69          1250000             82        NA    Combined Record            NA
#> 70          1250000             NA        NA  Government Record            NA
#> 71          1225000             NA        NA  Government Record            NA
#> 72          1220000             14        NA    Combined Record            NA
#> 73          1220000             NA        NA  Government Record            NA
#> 74          1211000             19        NA    Combined Record            NA
#> 75          1205000             NA        NA  Government Record            NA
#> 76          1200000             NA        NA  Government Record            NA
#> 77          1200000             NA        NA  Government Record            NA
#> 78          1183000             25        NA    Combined Record            NA
#> 79          1175000             NA        NA  Government Record            NA
#> 80          1165000             23        NA    Combined Record            NA
#> 81          1151200             NA        NA  Government Record            NA
#> 82          1126000             NA        NA  Government Record            NA
#> 83          1126000             23        NA    Allhomes Record            NA
#> 84          1125000             28        NA    Combined Record            NA
#> 85          1120000             NA        NA  Government Record            NA
#> 86          1100000             19        NA    Combined Record            NA
#> 87          1035000             NA        NA  Government Record            NA
#> 88          1025000             NA        NA  Government Record            NA
#> 89           950000             23        NA    Allhomes Record            NA
#> 90           950000             NA        NA  Government Record            NA
#> 91           900000             NA        NA  Government Record            NA
#> 92           890000             NA        NA  Government Record            NA
#> 93           735000             22        NA    Combined Record            NA
#> 94           720000             NA        NA  Government Record            NA
#> 95           635000             NA        NA  Government Record            NA
#> 96           575000             NA        NA  Government Record            NA
#> 97           450000             NA        NA  Government Record            NA
#> 98           250000            198        NA    Allhomes Record            NA
#> 99                0             NA        NA  Government Record            NA
#> 100               0             NA        NA  Government Record            NA
#> 101               0             NA        NA  Government Record            NA
#> 102               0             NA        NA  Government Record            NA
#> 103               0             NA        NA  Government Record            NA
#> 104               0             NA        NA  Government Record            NA
#> 105               0             NA        NA  Government Record            NA
#> 106               0             NA        NA  Government Record            NA
#>     land_type property_type                        purpose unimproved_value
#> 1          NA            NA    Single residential dwelling               NA
#> 2          NA            NA    Single residential dwelling           791000
#> 3          NA            NA    Single residential dwelling               NA
#> 4          NA            NA    Single residential dwelling           800000
#> 5          NA            NA    Single residential dwelling               NA
#> 6          NA            NA    Single residential dwelling               NA
#> 7          NA            NA    Single residential dwelling           824000
#> 8          NA            NA    Single residential dwelling               NA
#> 9          NA            NA    Single residential dwelling           860000
#> 10         NA            NA    Single residential dwelling           792000
#> 11         NA            NA    Single residential dwelling           637000
#> 12         NA            NA    Single residential dwelling           908000
#> 13         NA            NA    Single residential dwelling           622000
#> 14         NA            NA    Single residential dwelling           984000
#> 15         NA            NA    Single residential dwelling               NA
#> 16         NA            NA    Single residential dwelling               NA
#> 17         NA            NA    Single residential dwelling               NA
#> 18         NA            NA                           <NA>           622000
#> 19         NA            NA         Multiple unit dwelling               NA
#> 20         NA            NA                           <NA>               NA
#> 21         NA            NA    Single residential dwelling           949000
#> 22         NA            NA    Single residential dwelling           920000
#> 23         NA            NA    Single residential dwelling               NA
#> 24         NA            NA    Single residential dwelling           667000
#> 25         NA            NA    Single residential dwelling               NA
#> 26         NA            NA                           <NA>           693000
#> 27         NA            NA Single residential dwelling...               NA
#> 28         NA            NA Single residential dwelling...           851000
#> 29         NA            NA         Multiple unit dwelling               NA
#> 30         NA            NA    Single residential dwelling           832000
#> 31         NA            NA    Single residential dwelling               NA
#> 32         NA            NA    Single residential dwelling           873000
#> 33         NA            NA    Single residential dwelling               NA
#> 34         NA            NA    Single residential dwelling               NA
#> 35         NA            NA    Single residential dwelling               NA
#> 36         NA            NA    Single residential dwelling           873000
#> 37         NA            NA    Single residential dwelling           743000
#> 38         NA            NA Single residential dwelling...               NA
#> 39         NA            NA    Single residential dwelling               NA
#> 40         NA            NA    Single residential dwelling               NA
#> 41         NA            NA                           <NA>               NA
#> 42         NA            NA         Multiple unit dwelling               NA
#> 43         NA            NA    Single residential dwelling               NA
#> 44         NA            NA    Single residential dwelling               NA
#> 45         NA            NA    Single residential dwelling               NA
#> 46         NA            NA    Single residential dwelling           717000
#> 47         NA            NA    Single residential dwelling           792000
#> 48         NA            NA         Multiple unit dwelling               NA
#> 49         NA            NA    Single residential dwelling               NA
#> 50         NA            NA    Single residential dwelling               NA
#> 51         NA            NA    Single residential dwelling               NA
#> 52         NA            NA    Single residential dwelling           589000
#> 53         NA            NA    Single residential dwelling           664000
#> 54         NA            NA    Single residential dwelling           989000
#> 55         NA            NA                           <NA>           802000
#> 56         NA            NA         Multiple unit dwelling               NA
#> 57         NA            NA    Single residential dwelling               NA
#> 58         NA            NA         Multiple unit dwelling               NA
#> 59         NA            NA                           <NA>           951000
#> 60         NA            NA Single residential dwelling...               NA
#> 61         NA            NA         Multiple unit dwelling               NA
#> 62         NA            NA Single residential dwelling...           582000
#> 63         NA            NA Single residential dwelling...               NA
#> 64         NA            NA    Single residential dwelling               NA
#> 65         NA            NA Single residential dwelling...           636000
#> 66         NA            NA    Single residential dwelling           550000
#> 67         NA            NA Single residential dwelling...           791000
#> 68         NA            NA    Single residential dwelling               NA
#> 69         NA            NA    Single residential dwelling           824000
#> 70         NA            NA    Single residential dwelling               NA
#> 71         NA            NA    Single residential dwelling               NA
#> 72         NA            NA    Single residential dwelling           770000
#> 73         NA            NA    Single residential dwelling               NA
#> 74         NA            NA    Single residential dwelling           546000
#> 75         NA            NA    Single residential dwelling               NA
#> 76         NA            NA    Single residential dwelling               NA
#> 77         NA            NA    Single residential dwelling               NA
#> 78         NA            NA    Single residential dwelling           672000
#> 79         NA            NA    Single residential dwelling               NA
#> 80         NA            NA    Single residential dwelling           741000
#> 81         NA            NA    Single residential dwelling               NA
#> 82         NA            NA         Multiple unit dwelling               NA
#> 83         NA            NA                           <NA>           823000
#> 84         NA            NA    Single residential dwelling           770000
#> 85         NA            NA    Single residential dwelling               NA
#> 86         NA            NA Single residential dwelling...           550000
#> 87         NA            NA    Single residential dwelling               NA
#> 88         NA            NA         Multiple unit dwelling               NA
#> 89         NA            NA                           <NA>           676000
#> 90         NA            NA         Multiple unit dwelling               NA
#> 91         NA            NA    Single residential dwelling               NA
#> 92         NA            NA    Single residential dwelling               NA
#> 93         NA            NA         Multiple unit dwelling               NA
#> 94         NA            NA    Single residential dwelling               NA
#> 95         NA            NA Single residential dwelling...               NA
#> 96         NA            NA    Single residential dwelling               NA
#> 97         NA            NA Single residential dwelling...               NA
#> 98         NA            NA                           <NA>               NA
#> 99         NA            NA    Single residential dwelling               NA
#> 100        NA            NA    Single residential dwelling               NA
#> 101        NA            NA    Single residential dwelling               NA
#> 102        NA            NA    Single residential dwelling               NA
#> 103        NA            NA    Single residential dwelling               NA
#> 104        NA            NA         Multiple unit dwelling               NA
#> 105        NA            NA    Single residential dwelling               NA
#> 106        NA            NA    Single residential dwelling               NA
#>     unimproved_value_ratio
#> 1                       NA
#> 2                     0.32
#> 3                       NA
#> 4                     0.34
#> 5                       NA
#> 6                       NA
#> 7                     0.38
#> 8                       NA
#> 9                     0.41
#> 10                    0.40
#> 11                    0.33
#> 12                    0.47
#> 13                    0.33
#> 14                    0.53
#> 15                      NA
#> 16                      NA
#> 17                      NA
#> 18                    0.36
#> 19                      NA
#> 20                      NA
#> 21                    0.54
#> 22                    0.53
#> 23                      NA
#> 24                    0.39
#> 25                      NA
#> 26                    0.41
#> 27                      NA
#> 28                    0.50
#> 29                      NA
#> 30                    0.49
#> 31                      NA
#> 32                    0.52
#> 33                      NA
#> 34                      NA
#> 35                      NA
#> 36                    0.54
#> 37                    0.46
#> 38                      NA
#> 39                      NA
#> 40                      NA
#> 41                      NA
#> 42                      NA
#> 43                      NA
#> 44                      NA
#> 45                      NA
#> 46                    0.48
#> 47                    0.54
#> 48                      NA
#> 49                      NA
#> 50                      NA
#> 51                      NA
#> 52                    0.42
#> 53                    0.47
#> 54                    0.71
#> 55                    0.58
#> 56                      NA
#> 57                      NA
#> 58                      NA
#> 59                    0.71
#> 60                      NA
#> 61                      NA
#> 62                    0.44
#> 63                      NA
#> 64                      NA
#> 65                    0.49
#> 66                    0.42
#> 67                    0.62
#> 68                      NA
#> 69                    0.66
#> 70                      NA
#> 71                      NA
#> 72                    0.63
#> 73                      NA
#> 74                    0.45
#> 75                      NA
#> 76                      NA
#> 77                      NA
#> 78                    0.57
#> 79                      NA
#> 80                    0.64
#> 81                      NA
#> 82                      NA
#> 83                    0.73
#> 84                    0.68
#> 85                      NA
#> 86                    0.50
#> 87                      NA
#> 88                      NA
#> 89                    0.71
#> 90                      NA
#> 91                      NA
#> 92                      NA
#> 93                      NA
#> 94                      NA
#> 95                      NA
#> 96                      NA
#> 97                      NA
#> 98                      NA
#> 99                      NA
#> 100                     NA
#> 101                     NA
#> 102                     NA
#> 103                     NA
#> 104                     NA
#> 105                     NA
#> 106                     NA
# }